Initial commit
This repository implements base32 addresses for Bitcoin Cash.
The base32 proposal essentially copies the bech32 proposal by
Peter Wuille. Unfortunately, the proposal changes some of the
implementation features of the original specification making
this arguably the worst of the two.

- changes the checksum length (longer)
- base32 changes the field constants into 5 byte integers
- doesn't capture bits higher than 5th bit in the prefix in checksum computation
- current payloads can only deal with 160,192,224,256,320,384,448,512 bits in size
  instead of the arbitrary length sizes allowed by base32.
- base32 requires a biginteger library to run, doesn't
  work with native integer types in some languages
- uses ":" as the separator, so you have to click three times to
  select the whole line, or click and drag to select the whole
  address

but at least users will stop sending BCH to BTC and visa versa.